refactor user routes
